+++ chargebee-js-vue/src/components/Element.vue
@@ -24,6 +24,10 @@ export default {
         style: this.styles || {},
         placeholder: this.placeholder
       }
+    },
+
+    elementId: function() {
+      return `card-${this.id}`
     }
   },
 
@@ -41,7 +45,7 @@ export default {
     initializeField (cbComponent) {
       if (cbComponent) {
         const options = this.fieldOptions;
-        this.field = cbComponent.createField(this.id, options).at(`card-${this.id}`)
+        this.field = cbComponent.createField(this.id, options).at(`#${this.elementId}`)
         if (this.$parent.onMount) this.$parent.onMount()
 
         // Attach listeners if any