+++ chargebee-js-react/CHANGELOG.md
@@ -1,3 +1,7 @@
+### v0.1.3 (2019-05-09)
+* * *
+* Mount new syntax
+
 ### v0.1.1  (2019-04-28)
 * * *
 * Added Focus, blur, clear hooks